requirements.txt
flask
flask-cors
google-generativeai
py2neo
python-dotenv
gunicorn